Actually route to appropriate node
Despite the name, `route_to_node` was just choosing a random node from the ring
and using that for requests. This was fine for Ping and Noop operations, but for
timed_reads, we required the client to route operations to the node that owns
the key that it wants to read. This showed up on the results, as Antidote was
actually performing remote reads instead of reading locally.