Revert "No need to reprocess ring on socket creation"
Originally we were sending a list with duplicate IPs from Antidote, and failed
to take into account that the ring placement took into account this repetition,
so I changed Antidote to return a set instead. However, this broke the client
code partitioning. This revert makes it so we store the Ring untouched, but
still remove duplicates to prevent multiple open sockets to the same host.

This reverts commit 31ae34d3c2fcf30593484e9d97e2c26a85d77b13.